refactor: change the cjsToEsm file from commonjs to esm
