+++ src/modules/esm.mjs
@@ -1,0 +1,44 @@
+import path from "node:path";
+import { release, version } from "node:os";
+import { createServer as createServerHttp } from "node:http";
+import { fileURLToPath } from "node:url";
+import { dirname } from "node:path";
+
+// __filename ve __dirname oluşturma (ESM uyumlu)
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+// CommonJS dosyasını çalıştırır
+import "./files/c.cjs";
+
+const random = Math.random();
+
+// JSON'ları dinamik import ile alıyoruz
+const unknownModule =
+  random > 0.5
+    ? await import("./files/a.json", { with: { type: "json" } })
+    : await import("./files/b.json", { with: { type: "json" } });
+
+const unknownObject = unknownModule.default;
+
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const myServer = createServerHttp((_, res) => {
+  res.end("Request accepted");
+});
+
+const PORT = 3000;
+
+console.log(unknownObject);
+
+myServer.listen(PORT, () => {
+  console.log(`Server is listening on port ${PORT}`);
+  console.log("To terminate it, use Ctrl+C combination");
+});
+
+export { unknownObject, myServer };
